+++ services/director-v2/tests/unit/test_modules_dynamic_sidecar_volumes_resolver.py
@@ -1,8 +1,8 @@
 # pylint: disable=redefined-outer-name
 # pylint: disable=unused-argument
 
-from asyncio import FastChildWatcher
 import os
+from asyncio import FastChildWatcher
 from pathlib import Path
 from typing import Any, Callable, Dict, List
 from uuid import UUID
